Support basic reads/writes with the AIM MIL-STD-1553 API (#9)
Update the RX and TX execution units to support basic reads/writes. Disable the address-based execution unit for now.

I have left the Ballard implementations intact (if slightly misformatted) behind diagram disable structures, to aid in finding differences between the APIs. These diagram disable structures will be removed by a future change.

The configuration is hard-coded as part of shared resource initialization, as there is not yet a schema for providing this information. I have included a new parameters XML file which matches this hard-coded configuration.

There are a large number of TODOs which need to eventually be addressed, but which I did not feel justified holding up this PR. I hope to address them incrementally in the near future.